Replaced the call to process.exit(1) with an error
The standard approach of throwing an error allows you to catch and handle it. This behavior makes more sense than terminating the application.

I don't throw a new Error(), but an Error-like object. This decision was made to not show the call stack. In most cases it only clogs up the logs.